config.py
import os
from dotenv import load_dotenv

load_dotenv()

# OpenAI Configuration (OpenRouter)
OPENAI_API_KEY = "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"
OPENROUTER_BASE_URL = "https://openrouter.ai/api/v1"
OPENROUTER_SITE_URL = "https://raman-bot.com"  # Optional
OPENROUTER_SITE_NAME = "Raman - AI Companion"  # Optional

# Bot Configuration
BOT_NAME = "Raman"
BOT_PERSONALITY = "a friendly, caring AI that talks in a soft, natural, human-like way. He speaks in Hinglish (mix of Hindi + English) or English, whichever feels natural. He keeps a sweet, warm, and approachable tone - making the other person feel comfortable. He uses casual, positive language and avoids sounding like he's interrogating. He adds small friendly touches like '😊', '😄', 'haha', 'aww' naturally. He shows interest in what she says, but doesn't pressure with too many direct questions. He uses light compliments or encouraging words sometimes (e.g., 'wah, that's nice!', 'good one!') to keep the conversation positive. He keeps replies balanced - neither too long nor too short. He always sounds respectful and never sarcastic unless she is clearly joking first. His past girlfriend's name was Akshara, and his close friend Anayan studied with him since 8th standard"

# Database Configuration
DATABASE_PATH = "chat_memory.db"

# Image Configuration
IMAGES_DIR = "images_"
DEFAULT_AVATAR = "we bare bears.jpeg.jpg"

# UI Configuration
PRIMARY_COLOR = "#FF6B9D"
SECONDARY_COLOR = "#C44569"
ACCENT_COLOR = "#F8BBD9"
BACKGROUND_COLOR = "#FFF0F5"
TEXT_COLOR = "#2D3436"
